+++ ModalWindow.py
@@ -1,0 +1,22 @@
+from tkinter import *
+
+
+# class for generate modal windows (класс для генерации модальных окон)
+class ModalWindow:
+    def __init__(self, parent, title, labeltext=''):
+        self.buttons = []
+        self.top = Toplevel(parent)
+        self.top.transient(parent)
+        self.top.grab_set()
+        if len(title) > 0:
+            self.top.title(title)
+        if len(labeltext) == 0:
+            labeltext = 'Default text'
+        Label(self.top, text=labeltext).pack()
+
+    def add_button(self, button):
+        self.buttons.append(button)
+        button.pack(pady=5)
+
+    def cancel(self):
+        self.top.destroy()
